ecore-evas-wayland: Fix common window configure callback to not adjust for framespace
Summary: As we have already adjusted for framespace in various code
leading up to a configure callback, don't adjust for it here. This
fixes an issue where xdg surface window geometry would get incorrect
values which were including framespace. The values of the
xdg_surface_set_window_geometry should be Just the geometry of the
visible window.

@fix

Signed-off-by: Chris Michael <cp.michael@samsung.com>