Revert "Revert "Revert "evas: fix borked rotation support in GL engine."""
This reverts commit f51168f4ec7a1bd3072428c7a271ad7c6715fde5.

if this is right or not... this RESULTS in enlightenment becoming
totally unusable. put this back WHEN other issues are fixed you think
that exist. last i knew we query yinvert start with glx per native
surface and use that info - so as best i know it is and has been right
for a long time.

:)