hoversel: expands scrollable hoversel as large as it can
Summary:
In D2063, "max_size" data item is added to limit the number of items
to show at a time when hoversel is expanded.
However, it limits the number of items too few, so makes scrollable
function useless.
This patch removes limitation of hoversel size with pixels,
but if it needs, developers can set the limitation yet.
(By theme customization)

Test Plan:
elementary_test "hoversel"
click the second hoversel

Reviewers: DaveMDS, cedric

Reviewed By: cedric

Differential Revision: https://phab.enlightenment.org/D3223

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>