conform: if window is in background, indicator must not update the plug information.
Summary:
When the window is in background, for associated conformant, indicator must not update the plug information.
To fix this issue, On window goes to background, the associated plug handled by confromant is deleted.

Test Plan:
On any device (like handset) which should have indicator support, this scenario can be reproduce. I tested on a handset
with gdb attach to running process having window and confromant with indicator support.

Reviewers: raster, Hermet, cedric

Reviewed By: cedric

Subscribers: rajeshps, govi

Differential Revision: https://phab.enlightenment.org/D3177

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>