toolbar: reduce changing widget parent-child relationship when an item is added.
Summary:
It is an legacy of old code. When the view object of item was edje,
I think there was no meaningless parent-child relationship changes.
But, now, the view object is elm_layout and if we add object in proper order,
we don't need to make an useless parent-child realationship in any moments.

Test Plan: None

Reviewers: woohyun, cedric

Reviewed By: cedric

Differential Revision: https://phab.enlightenment.org/D3254

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>