segment_control: fix variable affectation
Summary:
When an Elm_Segment_Control object is created, the constructor calls
the smart_add() behind the scenes. It then calls _update_list() which
calls _position_items(), with the sd->obj unset (because it was set
after calling super constructors). This led to a CRI() (and therefore
to a backtrace) on the creation of each Elm_Segment_Control.

@fix

Test Plan: elementary_test

Reviewers: stefan_schmidt, cedric, raster

Differential Revision: https://phab.enlightenment.org/D3276

Signed-off-by: Cedric BAIL <cedric@osg.samsung.com>