evas-wayland-egl: Fix the unnecessary renewal of the engine window.
Summary:
As we do not set the ob->info->info.win variable anymore inside
Ecore_Evas (old code that was removed), we can omment out the line.
This line was causing unnecessary renewal of the engine window (thus
causing flickering when rotation was applied).

@fix

Test Plan: Rotate or Resize the window on wayland-egl.

Reviewers: gwanglim, devilhorns

Reviewed By: devilhorns

Subscribers: cedric, jpeg

Differential Revision: https://phab.enlightenment.org/D3407