Revert "evas: refactor model's savers and loaders."
This reverts commit 32c33ed64dda542c7cfc952fc656bb711260441b.

This refactor broke the evas test cases for the model loaders and savers. I gave
it a week to get fixed but a first try did not succeed and its blocks a lot of
other automated testing. To be honest, it should have never gone it when it
breaks existing test cases. Once fixed this refactor can happily go in.

Fixes T2905