Evas: Fix build for ubuntu... again
See 75fed54f0e4d72b20860
Apparently the proper CFLAGS are not set for SSE compilation.
SSE flags are set only for a specific file (now 2 of them).

As usual, it was worksforme but not for others. This should probably
fix that. Thanks Sub for the report and testing this solution.