elm config - allow the ability to save out to another hidden profile
support the ability to have hidden profiles that begin with a dot (.)
char like all hidden files/dirs. to do this we need to also be able to
create profiles with a given name (a hidden name) without switching to
them, so add the ability to save out to a specific profile name
without switching to it. of course this means to list profiles we need
to list them or list ALL includiing hidden profiles.

@feature