box: implement aspected item layout
tl;dr: if you haven't used aspect hints on objects packed into boxes,
this commit has no effect on anything you've done

in some cases, using weight+align hints are not sufficient in order to
do a simple box packing of objects.
one example of this is when a box's size is not explicitly determined,
such as if it is packed with EXPAND into another container and contains
only objects with no min size hint, such as images. in this case, the box
will determine that its min size is extremely small, resulting in nearly-hidden
contents inside the box.

using aspect hints, this outcome is noticeably different. the expanded/filled
potential size of a box item is used to calculate the aspected size, and this
size, if it fits within min/max constraints, is then used to calculate the min
size of the box.

as a more concrete example:
* create box with maximum size 100x50
* pack 2x image into box
 - set VERTICAL aspect, EXPAND, FILL
have 2x 50x50 images

also I added some docs

@feature