elm config - add the ability to have derived configurations from others
this adds the ability to have one profile automatically be derived
from another with modifications applied. right now it only can do a
very limited modification - multiple scale. over time this will
expand. this is only inteded otbe sued by a DE like enlightenment, so
it's not going to be documented fully at this point.

@feature