Elm entry: Fix selection using handlers.
Summary:
Fix item selection using selection handler

In case of selection using handler, there is a problem in case the entry
has the item, mouse down over handler consider the current down
coordinate to start selection and it becomes very difficult to select in the
situation shown in the test plan. So I consider the next/previous  line
geometry to fix it.

@fix

Test Plan:
1.Enable selection handler
2. Have a look at the setup in the attached image
{F27396}

3. Start selecting from last item and extent to cover the last text.
4.Observe that its impossible to select the last text until you move mouse out of the window
So, in mobile device its almost impossible to select using handlers in this situation.

Please use the attached application to test it.
{F27402}

Reviewers: thiepha, herdsman, tasn

Reviewed By: tasn

Subscribers: shilpasingh

Differential Revision: https://phab.enlightenment.org/D3479