Gengrid: improve item_region_show code
There is no need to recalc row and col position of the item, they are already stored in the item struct.

The old implementation was calculating wrong values and also was storing this wrong values in it->x and it->y, resulting in wrong results for the elm_gengrid_item_pos_get() function.

so, at the end,  this is a @fix for the pos_get() function.

Also changed a bit the test to let the gengrid fill the window and thus testing behaviour on col/row changes.

I didn't find any regression in all the gengrid tests after this, let me know if it broke something for you.