Evas GL: Actually save the checksum inside the file
This would be cleaner than adding it as a suffix to the filename.
Use simple write and read_direct for simplicity.

This is an improvement on 3dcd903a537e29342121c and should make sure
everyone is actually using the latest version of the shaders even
if they track master or the alpha/beta releases.