Add a test to reveal a focus bug in Genlist/Gengrid
To reproduce:
 * elm_test -to "GenGrid Focus"
 * turn on the "Focus Highlight" checkbox
 * double-click an item to show a popup that should have the focus

As you can see the mouse-up event on the gengrid item steal the focus from the popup.
When, instead, the item is activated with the keyboard the popup correctly keep the focus.

I have this problem in both genlist and gengrid, so probably the error is not in the gengrid code, but somewhere else that I cannot find.

...any help appreciated