Genlist: do not select the item two times in case of a double click.
In the case of a genlist set as ELM_OBJECT_SELECT_MODE_ALWAYS the "selected" cb was called 2 times. This @fix that by skipping the selection code in case the click is not the first one (EVAS_BUTTON_NONE).

Also merge the conditions of two consecutive "if" in a single statement.